Initial start for docker/flask dev
