--- config.json
@@ -1,14 +1,0 @@
-{
-	"slack":{
-		"SLACK_BOT_TOKEN" : ""
-	},
-	"misp" : {
-		"url" : "", #must include trailing /
-		"ssl" : true,
-		"key" : ""
-	},
-	"logging" :{
-		"output_file" : "directory/for/log.log",
-		"output_error_file": "directory/for/error.log.log"
-	}
-}
