--- machinetag.json
@@ -1,33 +1,0 @@
-{
-  "predicates": [
-    {
-      "colour": "#CC0033",
-      "description": "Observed phishing content",
-      "value": "phish"
-    },
-    {
-      "colour": "#FFC000",
-      "description": "Observed malware",
-      "value": "malware"
-    },
-    {
-      "colour": "#339900",
-      "description": "Observed BEC/SCAM",
-      "value": "bec/scam"
-    },
-    {
-      "colour": "#ffffff",
-      "description": "Information from a data dump.",
-      "value": "dump"
-    },
-    {
-      "colour": "#d208f4",
-      "description": "Believed to be possible APT related indicators",
-      "value": "apt"
-    }
-  ],
-  "version": 1,
-  "description": "Tags Defined for IR8 specific information",
-  "namespace": "ir8",
-  "exclusive": true
-}
